feat: add list tasks command

+++ src/main/java/models/Chore.java
@@ -8,7 +8,7 @@ import java.util.List;
 @Data
 public class Chore {
     @JsonProperty("assigned_ids")
-    private List<Integer> assignedIds;
+    private List<Long> assignedIds;
     @JsonProperty("assigned_usernames")
     private List<String> assignedUsernames;
     private Boolean done;
